+++ src/components/LecturePages/LectureDetailPage/RateBarChart/index.tsx
