+++ src/components/Shared/RateDisplay/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Star from '@components/Shared/RateDisplay/Star';
+import Star from './Star';
 
 const Container = styled.div`
   display: flex;
