--- src/pages/lectures/[id].ts
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@views/LecturePages/LectureDetailPages/LectureDetailPage';
