+++ src/pages/lectures/[id].tsx
@@ -1,0 +1,1 @@
+export { default, getServerSideProps } from '@views/LecturePages/LectureDetailPage';
