+++ src/views/LecturePages/LectureDetailPage/LectureDetailPage.style.ts
